requirements.txt
zeep
redis
bigcommerce
phonenumbers
php
pytest==6.2.4
requests_oauthlib==1.3.1
sentry-sdk==1.6.0

